IMPALA-12243: Add support for DROP PARTITION for Iceberg tables
This change adds support for DROP PARTITION operation for Iceberg
tables. Users can now execute 'ALTER TABLE table DROP PARTITION
(partition_filtering_expression)' statements where
'partition_filtering_expression' consists of at least one binary
predicate, IN predicate, IS NULL predicate or arbitrary logical
combination of these, where the predicate's term is a partitioned column
decorated with a partition transform, for example: 'day(column)' or
simply column.

To select partitions with mutating transforms (non-identity transforms),
the user must provide the transform in the selection, for example: if
column 'a' is partitioned as year(a), the user must filter it with
explicitly providing the transform, like (year(a) = "2012"). This is a
requirement for Iceberg's planFiles API. If the column is filtered
without the transform, the filtering of datafiles with strict projection
results in ambiguous results.

The binary predicate's operand can be basic comparison operators: =, !=,
<, >, <=, >=.
The IN and IS NULL predicates can be inverted with the NOT keyword.

Only constant literals are accepted currently for these predicates; for
example: (identity_string = "string") is accepted, (identity_string =
concat(identity_string, "another_string") is not accepted.

Tests:
 - unit tests for IcebergUtil.getDateTransformValue
 - analyzer tests for negative cases, simple cases
 - e2e tests for every transform type, schema evolution and rollback

Change-Id: I2a768ba2966f570454687e02e4e6d67df46741f9